event-based-funcs/recommend/requirements.txt
# Function dependencies, for example:
# package>=version


google-cloud-firestore==1.1.1
google-cloud-pubsub==0.40.0
Flask==1.0.2
